Example weapon merge (ExampleYoyo)
