+++ ExampleMod/Items/Weapons/ExampleYoyoProjectile.png
